+++ include/bytecode.h
@@ -52,7 +52,9 @@ enum {
     BC_GT,
     BC_GTE,
     BC_AND,
-    BC_OR
+    BC_OR,
+    BC_PUSH_CHAR_RAW,
+    BC_PUSH_STRING_RAW
 };
 
 void bc_init(Bytecode *bc);